backend/movie-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo==4.6.1
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiokafka==0.8.1
python-dotenv==1.0.0 